% Auto-generated: do not edit by hand
\name{aceEditor}

\alias{aceEditor}

\title{AceEditor component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
aceEditor(id=NULL, label=NULL, value=NULL, className=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{value}{Character. The value displayed in the input.}

\item{className}{Character. Often used with CSS to style elements with common properties.}
}
