% Auto-generated: do not edit by hand
\name{dashAceEditor}

\alias{dashAceEditor}

\title{DashAceEditor component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashAceEditor(id=NULL, value=NULL, className=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{value}{Character. The value displayed in the input.}

\item{className}{Character. Often used with CSS to style elements with common properties.}
}
